refactor: separate navigation component
