+++ src/lib/layout/navigation/navigation.module.css
@@ -1,0 +1,87 @@
+.navigation {
+	--icon-size: 1.25em;
+
+	position: sticky;
+	top: 0;
+	z-index: 10;
+	display: grid;
+	grid-template: 1fr / 1fr;
+	background-color: var(--color-surface-0);
+	border-bottom: 3px solid hsl(0deg 0% 50% / 15%);
+	margin-bottom: var(--space-8);
+	font-size: var(--font-size-2);
+	width: 100%;
+	padding-inline: var(--space-2);
+}
+
+.navigation > form {
+	display: none;
+}
+
+.navigation ul {
+	list-style: none;
+	justify-self: center;
+	display: flex;
+	align-items: center;
+	gap: var(--space-4);
+	height: var(--tabs-height);
+	padding-block: var(--space-3);
+	margin-inline: auto;
+	width: var(--max-width);
+
+	overflow-x: auto;
+	scroll-snap-type: x mandatory;
+	scroll-padding-inline-start: var(--space-4);
+}
+
+.link {
+	position: relative;
+	scroll-snap-align: left;
+	color: inherit;
+	display: flex;
+	place-content: center;
+	gap: 0.5em;
+	background-color: var(--color-surface-2);
+	padding: var(--space-2);
+	border-radius: var(--radius);
+	white-space: nowrap;
+	user-select: none;
+	text-decoration: unset;
+	transition: background-color 0.125s linear;
+}
+
+.link:hover {
+	text-decoration: unset;
+	background-color: var(--color-surface-3);
+}
+
+.link:active {
+	background-color: var(--color-surface-2);
+}
+
+.link[aria-current] {
+	background-color: var(--color-accent);
+}
+
+.link:focus-visible {
+	outline: 2px dashed var(--color-accent);
+	outline-offset: 2px;
+}
+
+@media screen and (width <= 600px) {
+	.navigation {
+		position: fixed;
+		top: unset;
+		bottom: 0;
+		left: 0;
+		height: min-content;
+		margin: 0;
+		border-block-start: 2px solid var(--color-surface-2);
+		border-block-end: unset;
+		padding-inline: 0;
+	}
+
+	.navigation ul {
+		padding-inline: var(--space-4);
+	}
+}
